change mask to int1 datatype, real md5sum ok

+++ src/General/memory.f90
@@ -112,7 +112,7 @@
       INTEGER, allocatable :: mbathy(:,:)
 
 
-      double precision, allocatable, dimension(:,:,:) :: tmask,umask, vmask
+      INTEGER(kind = 1), allocatable, dimension(:,:,:) :: tmask,umask, vmask
       INTEGER NBFMPOINTS, NBFMPOINTS_SUP, NWATERPOINTS
       INTEGER, allocatable, dimension(:,:) :: BFMpoints
